CB-26322 Do not set cdpHashedPassword attribute if cdpUnencryptedKrbPrincipalKey missing
When setting a users password, we only check if cdpHashedPassword is present.
This could cause an issue in FreeIPA if cdpUnencryptedKrbPrincipalKey is not present.
To make sure we have both present, a check needs to be added.
It would help investigate issues where we previously seens this:

```Invoke FreeIPA failed: Operations error: must specify both cdpUnencryptedKrbPrincipalKey and cdpHashedPassword if either is specified```

If after this change we see the same issue, then the information is lost on FreeIPA side.