CB-26328: Fix freeipa upscale failing with "Access to KMS is not allowed" when secret encryption enabled
When we are creating the userdata secrets, by mistake, we were setting the KMS key of the secrets to the actual KMS key to be used with them later. The problem with this, is that we update the policy of the KMS key to restrict access to it to the already created secrets, and when we upscale the stack, the new secrets can't be created, because we are not allowed to use the KMS key, as the new secrets haven't been added to the key's resource policy yet.
The fix is to use the default AWS managed key when creating the secret, and only later update the key to the actual key of the stack.