CB-26401 Implement logic to allow/hide secret types based on Public Cloud deployment
Filter enabled when:
- fips mode disabled
- secretrotation.commercialEnabledSecretTypes is not empty

If the filter is enabled and fips mode is disabled, the list endpoint will return the configured secret types only. It will also validate the rotation request.

The application will not start if an invalid secret type is configured (unknown or internal secret types are not allowed)

Secret type descriptions updated.