CB-26192 Add 'trust' to the preferred time source for chrony
In chrony it could happen that 2 sources are present, as we ship our images with the default config and add the cloud provider's time source.
This could result in an issue, when chrony couldn't decide which source is right, which could result in blocking time sync.

So adding `trust` to the cloud providers source it ensures chrony will always prefer it, as long as it's available.

The salt state is change to replace, to ensure we update the existing one instead of adding a new one to the config.