CB-26778 Add retry to Azure database deployment
In case a deployment or it's output is null, a retry is added. This is to avoid NullPointerException as output always should be present.
This seems to be an Azure bug that the deployment doesn't have the output, but without it we cannot proceed, as FQDN is required.
The original logic is kept, so if there was an exception earlier, it would be thrown, while if there was no exception, but the deployment or output null,
then a `ActionFailedException` would be thrown. This is required for the retry.