CB-26690 Provisioning for FreeIPA cluster on GCP should support Multi AZ Provisioning for FreeIPA cluster on GCP should support Multi AZ
For GCP, Subnet is not a zonal service but Az is mandatory field so multiAz flag decides if it is Fixed Az or instances are distributed across multiple AZs. Current implementation is supporting Fixed Az by picking random Az, populate it at subnet level and it works the same way as AWS.

To support MultiAz on GCP, We need to ensure that Subnet's Az is only populated at instance level if multi Az is false otherwise We need to execute additional step to populate AZ for instances like Azure.

For GCP, if multiAz is false , it will use the logic same as AWS for populating Az on instances. if multiAz is true, it will use the same logic as Azure for populating Az on instances.