CB-27046 Downscale put stack into DOWNSCALE_REQUESTED before flow is triggered
Convert DOWNSCALE_REQUESTED and UPSCALE_REQUESTED detailed statuses as an AVAILABLE status. This change is required to prevent errors when the cluster is stuck in the IN_PROGRESS status if the flow couldnt start. These statuses are required in order to have proper status reason for EDH sending (the EDH event sending happens before the first flow state)