CB-27150 Stop FreeIPA services before instance shutdown
Sometimes dse.ldif got corrupted or missing. The first guess is systemd doesn't wait for the processes to finish and terminates them causing the issue.
With this we explicitly stop the services before stopping the instances. The delay is moved to stop the services and instances stopped at the same time.