CB-27167 Azure flexible db upgrade workaround: check db connection before restoring the backed up data
Azure flexible db upgrade workaround: check db connection before restoring the backed up data.
It is necessary as upgrade to flexible server is blocked because of CB-27094, and a workaround is needed.
The workaround is to check the database connection via a salt state for 1hour. If the connection is ok, then we continue with the restore step. If connection is failed for an hour, upgrade flow will be failed.