CB-27134 FreeIPA rebuild - prepare for old AzCopy version
azcopy version in 2021 did not have a parameter (--as-subdir=false) that we use.
So instead of using this flag, it's omitted and `/*` is appended to the URL so only the content of the folder is downloaded. The result is the same.

Tested with an old image which is affected, and with the most recent one too.