CB-27120 Retry should work when the RPM download fails
There is a check which validates if upgrading the RPMs are required or not.
If some of the instances managed to update the RPMs while others not, this check threw an error about the mixed versions.
From now on, if we can't decide if the upgrade is required, the salt states will always run.