CB-26364 Deprecate SharedService response in Datahub list operation to fix service slowness
The service slowness was caused by iterating over every Datahub (e.g., 86 entries on mow-dev) and making callouts to CDL and Datalake just to determine CRN and name. This data was being added to the now-irrelevant SharedService field of the Datahub response, unused since 2019.

Instead of optimizing the retrieval, I deprecated the SharedService response and left it empty. Although this is a breaking change, a search in github.ifra showed no services relying on this field, so the risk is minimal.