CB-26981 Optimize UmsUsersStateProvider's WorkloadAdministrationGroup hanlding
Instead of creating a separate `ImmutableSet` for every environment from `WorkloadAdministrationGroup`,
only one is created.
The benefit is that it's very expensive for a large number of WAGs and a quite several environments to create these sets.
Set creation invokes a lot of hash and equal check, while the result is the same, so we can reuse a single one.