CB-27355 Proceed with node replacement only after the CM reports that all services are healthy during a HA OS upgrade
During the upscale, the hostsStartRoles command finishes successfully regardless of the HDFS health status. This can lead to HDFS outage. To improve this, an additional poller was added to wait until the CM services become healthy. The max wait time is 5 minutes. Only enabled if the rollingRestartEnabled flag is true (HA OS upgrade).