CB-27249: Add new validation for DataLake resize on Azure to allow resize operation only with Flexible database servers
In this commit, I've added a new validation to DataLake resize on Azure to disable the resize operation if the customer's cluster using Azure Single Server.
The porpuse of this validation is that the Azure Single Server is deprecated therefore before the resize operation the customer must upgrade the database to Azure Flexible Server first.