CB-27057 Sign key format not compatible with Salt 3006.9
Use PEM format for salt and saltboot public keys instead of the OpenSSH format and implement salt sign key pair rotation in DL and DH
Private keys are stored in Vault while the public keys are stored in the database. Due to how the secret rotation framework works, the vault rotation step will change the private key in the Vault and the custom_job rotation step will generate the new public key in the database based on the private key. In case of a rollback the steps are very similar, the public key part gets generated based on the original private key in Vault.