CB-27422: Block upgrade path to 7.3.1.0 from specific runtime versions
In this commit, I've introduced a new validation to prevent the customers to upgrade to 7.3.1.0 from specific versions.
We're blocking the upgrade if the current cluster runtime version is:
- 7.2.17.100 and lower
- 7.2.17.600 and higher
- 7.2.18.300 and higher

This commit contains a specific parameterized unit test to cover all combination
I also extended the upgrade component test with this case.