CB-24646 Freeipa salt masters have different private keys in HA applications
Private keys are stored in Vault while the public keys are stored in the database. Due to how the secret rotation framework works, the vault rotation step will change the private key in the Vault and the custom_job rotation step will generate the new public key in the database based on the private key. In case of a rollback the steps are very similar, the public key part gets generated based on the original private key in Vault.

If the related fields in the database and Vault are empty (clusters created before this PR), generate and store a new keypair as a "base". This means, that in case of an error, we cannot rollback to the previous version (exists only on the cluster) only to this base.