CB-27578 - Flexible Server Pg 11 upgrade to Flexible Server Pg 14
This commit implements the in-place Flexible -> Flexible upgrade.
My tests have run for around  ~17-25 minutes.
According to MS docs (https://learn.microsoft.com/en-us/azure/postgresql/flexible-server/concepts-major-version-upgrade#process), it happens like the following:

"During the process of an in-place major version upgrade, Azure Database for PostgreSQL flexible server runs a pre-check procedure to identify any potential issues that might cause the upgrade to fail.

If the pre-check finds any incompatibilities, it creates a log event that shows that the upgrade pre-check failed, along with an error message.

If the pre-check is successful, Azure Database for PostgreSQL flexible server stops the service and takes an implicit backup just before starting the upgrade. The service can use this backup to restore the database instance to its previous version if there's an upgrade error."

Tested with
- Private setup
- CMK setup
- Delegated subnet setup