CB-27629 Reject the DB upgrade request on Azure if the CDP_POSTGRES_UPGRADE_SKIP_SERVICE_STOP is granted
We have an entitlement to skip the service stop/start and it's called: CDP_POSTGRES_UPGRADE_SKIP_SERVICE_STOP. This has worked in the past when we could keep the hostname user/pass for the database, but it is not the case for Flexible server. Given this we should not allow any DB upgrade on Azure from Single server to Flexible server when this entitlement is enabled. We should send a notification message to the customer that this entitlement is enabled and open a Cloudera support ticket to get it revoked.