CB-24859 fix network entity related duplicate key value violation issue
The issue arises due to two factors:

@Where(clause = "archived = false") annotation on BaseNetwork
https://github.infra.cloudera.com/cloudbreak/cloudbreak/blob/89af7fa79ed954f6687d511b7a5f739c94af4ba4/environment/src/main/java/com/sequenceiq/environment/network/dao/domain/BaseNetwork.java#L47
environment_network_environment_id_idx, which is a unique index on environment_id in the environment_network table.
How can this occur?

Fact: two or more operations can happen simultaneously in the same environment:

A create operation and a termination operation (the termination process can begin even if another flow is currently running).
Multiple termination operations can be executed at the same time.
A combination of both types of operations can take place.
A flow fetches the environment and sets a new status for the environment, like here:

https://github.infra.cloudera.com/cloudbreak/cloudbreak/blob/71034f51ab76e02a02e6a19c59628c68d7eef133/environment/src/main/java/com/sequenceiq/environment/environment/service/EnvironmentStatusUpdateService.java#L43-L54

Take a note the find and the save are not in a transaction:

environmentService.findEnvironmentById reads the environment and there is a network for the environment
environmentService.save(environment) -> saves the environment with network info

Here's the trick: If an another running termination flow sets the network to archived at  https://github.infra.cloudera.com/cloudbreak/cloudbreak/blob/6a3884806736ef8786914559e4e6b693a33dc8e7/environment/src/main/java/com/sequenceiq/environment/environment/flow/deletion/handler/NetworkDeleteHandler.java#L67
then when `environmentService.save(environment)` is executed, JPA will notice that there is no active network for the environment (due to our beloved @Where and the clause "archived = false"). As a result, JPA will try to save a new network into the DB. However, we already have an archived network stored in the database so the unique index will fail.

Mitigations in this pull request:

- When deleting a network resource, avoid using the previously retrieved environment object, as it may be outdated.
- Fetch the environment and archive the network within a transaction during the deletion of a network resource.
- Instead of saving the entire environment entity during an environment status update, implement a new repository call specifically for the environment status update.