CB-27714 - Azure SDK update broke whatIf analysis
WhatIf analysis code works with azureSdkResourceManagerVersion 2.40.0 but breaks with version 2.42.0. Further investigation revealed that this behavior is due to version 2.42.0 switching to azure-json for serialization/deserialization. MS support verified this behavior by removing the azureSdkResourceManager version, and the script executed successfully.

MS wants to highlight that "JsonNode" is from Jackson, and currently, they do not support directly passing it in as a property.  As discussed, please try replacing JsonNode with Map in the code below and proceed with the deployment.  Reference:  azure-sdk-for-java/sdk/resourcemanager/azure-resourcemanager-resources/CHANGELOG.md at main · Azure/azure-sdk-for-java · GitHub