CB-27760 Fail when reverse DNS record creation is not successful
The `add_dns_record` script didn't fail, if it hasn't founnd a reverse zone to create the record in.
Now it would fail.
Also the log has been adjusted to push only the needed info to stderr so we can display it in the statusreason.

In env service the FreeIPA creation handler is adjusted so it would create reverse zones if the pod is restarted while waiting for FreeIPA creation.