CB-27918 Refresh Cluster command fails post (COD ZDU or EDL rolling) OS upgrade
FULL_DOWNSCALE_TRIGGER_EVENT is triggered to execute hostsStopRoles command during the downscale phase of the OS upgrade even if the primary gateway is in the actual upgradeSet, because CM server should be always available during the OS upgrade.
Before this change, the services were not stopped during the node replacement (in case of the upgrade set with the primary gateway). As a result, during the upscale phase, services started with outdated configs causing unhealthy hdfs and hbase statuses and also the failure of the refreshCluster command.

Affected flows:
- EDL rolling OS upgrade
- COD ZDU OS upgrade