CB-28068: Switched `swagger-compatibility-check` library to `tufin/oasdiff`
Pretty much a drop in replacement, with many advantages:
- More verbose output, including exact names of compatibility breaking fields
- Significantly more configurability options, like being able to define what severity should each change be treated as (error/warning/info)
- Open source, under Apache 2.0 license (same as current tool), actively developed

The current tool we use works, but when there is a compatibility problem, it will tell you the endpoint that changed, but you will have to navigate through all the nested classes to find the exact field that is causing the issue. Also the range of configurability options is quite narrow, and whether a type of api change should break the compatibility is a boolean option instead of being one of error/warning/info, like in the other tool.