CB-28119 Increase retry for FreeIPA rebuild `update_dns_record_rebuild.sh`
On Azure it takes more time for the FreeIPA services to stabalize than the current settings.
While both `ipactl` and `systemctl` returns with and everything is up&running state, some of the services are still not operable fully.
This resulted in a loop where a new restore would be started and result in the same issue. Increasing retry gives the system enough time to stabalize after restore.