CB-27786 Run OS upgrade automatically after the runtime upgrade
Add new entitlement: CDP_DATAHUB_FORCE_OS_UPGRADE

The replaceVms field is currently used only in tests (the UI and CLI leave it empty). Its default value is true. If the CDP_DATAHUB_FORCE_OS_UPGRADE entitlement is disabled, the backend may change its value to false if lockComponent is false (indicating a runtime upgrade without an OS upgrade). If the entitlement is enabled, the backend will not change its value, allowing the runtime and OS upgrade to be executed together - if other requirements are met (no ephemeral volumes and gateway repair enabled).

If the entitlement is enabled, rollingUpgrade flag is true and it is a COD cluster OR the cluster size is larger than the allowed node count for rolling upgrade (cb.upgrade.validation.distrox.maxNumberOfInstancesForRollingUpgrade), the OS upgrade (repair) will be executed on the Salt master nodes only.