CB-28147: Turn off update proxy config E2E test, as it is not supported for AWS_NATIVE currently
The current implementation requires a userdata update and also a salt state update. The problem is, that for non Cloud Formation EC2 instances, you can only update the userdata if the instance is stopped. However, the instance needs to be running for the salt state update. So the current implementation would need to be significantly changed for proxy config update to work on AWS_NATIVE.