CB-28186 Use default yarn.scheduler.capacity.resource-calculator
DefaultResourceCalculator was set for most of the blueprints, but the DominantResourceCalculator (which will be set as default) offers more flexibility, like requesting multiple vcores for each executor.
Quoting from CM description of the property: `The Resource Calculator implementation to be used to compare Resources in the scheduler. The DefaultResourceCalculator only uses Memory while DominantResourceCalculator uses Dominant-resource to compare multi-dimensional resources such as Memory, CPU etc.`