CB-28261: Remove unused fields with `@Inject`
Removed all unused fields found by the `UnusedInjectChecker`.

The `UnusedInjectChecker` util can find unused fields which have the `@Inject` annotation, however we cannot let it run as a regular unit test, because the `MemberUsageScanner` used doesn't always find all the usages, missing a couple of them roughly every 5th run of a test, making the tests using it be flaky because of the false positives. This util will never make the tests using it fail, but it will still log if there are unused fields found (even if they are false positives).

The dependecy providing the scanner (`org.reflections:reflections`) has not been updated since October 2021.