CB-28346 Trust CMCA cert when deploying cluster with AD
The CM agent copies the certificates to itself after it start up the first time and connects to CM.
There is a wait as usually agents started sooner as CM. Once the cert is there, it's imported.
A check is in place to compare fingerprints and avoid cert import on every highstate.