CB-28369: Handle heterogeneous saltboot HTTPS settings between nodes
When we are collecting and checking the fingerprints received from the saltboots, we are comparing the received address against the expected IP address, but when not all saltboots are listening on the same port, the addresses in the `FingerprintsResponse` also contain the port the saltboot responded on. During upgrades it is expected that some saltboots are communicating on port `7070` and some are on `7071`.

Added the method `getIpFromAddress` to get only the IP address from the `address` field even if the port is also present.