CB-28444: [E2E] Add extra check to ensure secret encryption is enabled for the environment before trying to validate it
Previously it was assumed, that on AWS_GOV secret encryption is always enabled in the E2E tests. This should hold true in most cases, but if it has to be turned off for some reason, the tests should still be able to run without failing just because the secret encryption assertions fail.