CB-28347 Salt version upgrade on FreeIPA clusters
Use the gateway with the latest salt version (if exists) instead of the primary gateway during the Salt communication in FMS. This is required, because a salt minion with a newer version cannot connect to an old salt master, so we have to ensure to always connect to the gateway with the latest salt version. Unfortunately, we cannot use the ChangePrimaryGateway flow here, because the stack and cluster creation is one flow in FMS. This solution was suggested and approved by @mmolnar.

Upgrade steps:
1. change image
2. provision a new instance using the new image - temporarily use this instance instead of the PGW to communicate with the salt cluster to successfully finish the provision
3. change PGW to the newly created instance
4. downscale the old PGW
5. repair the remaining nodes

Additional changes:
- Save salt version in the DB
- Reorder repair flows during upgrade to upgrade the PGW first