CB-28205: Offset devices names on detached `AWS_VOLUMESET` resources by the number of ephemeral volumes
On AWS_NATIVE, when you vertical scale an instance to an instance type which has a higher number of ephemeral volumes and then repair the scaled instances, the attachment of our attached volumes will fail, because AWS mounts the new ephemeral volumes under the same device names we have saved as the device names for our attached volumes causing a collision.

To solve this, we offset the device names in our existing `AWS_VOLUMESET` resource by the number of ephemeral volumes the new instance types has in total. This is the same thing we have already been doing during creation.