CB-28429 Extend the FreeIPA downscale flow with Load Balancer related logic
Instead of removing the downscale candidate instance from the LB backend pool, the health agent on the instance(s) are stopped.
After the stop, there is a 30sec wait to give enough time for the LB to realize it's down, and it should redirect the requests to another instance.

Only happens if there is an LB present.
The deletion of the instance on the cloud provider side implicitly removes it from the LB's backend pool.

If the stop fails, the downscale process will continue anyway.