CB-28837: Reject Basic SKU option for Loadbalancer
In this commit I've created a validation for Azure Load valancer SKU.
From now we'll throw a BadRequestException for the cluster creation request if it contains BASIC SKU for load balancers.
The only accepted values are the null and the STANDARD SKU
This change is affects the DataHub and DataLake clusters as well.