CB-29219 Fix Freeipa fallback salt password rotation
Without this the salt password that we are sending to salt-bootstrap will remain the old one, it will re-create the user with the old password, and then the new password (which was not used anywhere) will be saved to vault.
Also added a validation after the rotation to make sure we are not saving wrong password for salt.