CB-29238: Activate only non-cdh parcels before runtime upgrade command
In this commit, I've fixed a bug in the runtime upgrade process when we're upgrading non-cdh services as well.
The problem was that if we tried to upgrade a cluster with FLINK and CDH then we downloaded, distributed and activated both services but only the FLINK (non-cdh) service needs to be activated.
And therefore the upgrade CDH command failed because the new CDH version was already activated.
In this fix I've modified the parcel activation part to activate only the non-cdh services.