CB-29348 Update heartbeat on application startup
Heartbeat is running every minute 0 and 30 seconds. If the app is started eg at minute 05 seconds, then there is a 25 sec gap when the flow engine would reject requests with {{Current node is not present in the database!}} from com.sequenceiq.cloudbreak.ha.service.NodeValidator

The solution is to update heartbeat on startup.