CB-29462 AWS native Loadbalancer resource name can collide
If the environment/stack first part used in the naming is the same and we start enough environments or clusters there could be a name collision for loadbalancers, similarly to targetgroups:
https://jira.cloudera.com/browse/CB-26486

The solution is to replace timestamp with part of the CRN, also shorten the middle part `LBInternal` to `Internal` by skipping the addition of `LB`. This would give us 2 extra character for the stack name.
Although we have to keep the scheme time, as we filter loadbalancers based on it during scaling, and it could not be dropped easily as attributes contain this information only after 2023, while the feature was introduced in 2021.