CB-29400: The roll-forward image sync is updating to a newer image even when it's not required
In this commit I've modified the image sync part of the roll-forward to use the current image if available as a candidate.
Previously we had an issue where after a failed upgrade the syncer changed the image to the newer one, because the newer image also contains the same component versions as the current image.
From now we're selecting the current image if available as a candidate.