CB-29413: Refactor metering instance type checker to also check for orphan instances
The `MeteringInstanceChecker` has been renamed to `InstanceChecker` in most places.

The job is scheduled, if either
1. `instancechecker.enabled` is `true`
or
2. both `meteringingestion.enabled` and `meteringingestion.instanceCheckerEnabled` are `true`.

The job itself will only perform orphan resource detection if the 1. point is true, and only sync the instance types if the 2. point is true. The two operations can be enabled at once.
The job was also changed to run for both DataLakes and DataHubs, but the instance type sync will only be triggered if the stack is a DataHub.

The orphan resource detection collects all the instances on the provider which are not deleted, and have the tag with the key `Cloudera-Resource-Name` (or `cloudera-resource-name` in case of GCP) and value equal to CRN of the stack. Then it compares this list of instances to the list coming from the `instancemetadata` table, and looks for instances which are present on the provider, but their instanceId is not present in the `instancemetadata` table. If it finds instances like this, it will send a notification with the specific instanceIds to the customer.