CB-29434: Fix rolling upgrade validation for DataLake clusters
In this commit, I've fixed an uncovered scenario when the users can perform a rolling upgrade with LIGHT_DUTY template when the CDP_SKIP_ROLLING_UPGRADE_VALIDATION entitlement is enabled.
This can cause upgrade failure in the CM because the non-HA CM is not able to process the rolling CDH upgrade request.

I also performed some refactoring because it was too complicated to write unit tests for the affected validation logic therefore I've extracted into a separate class.