CB-29520 - Cluster upgrade validation skipped. Reason: Cannot invoke "com.sequenceiq.cloudbreak.api.endpoint.v4.stacks.request.database.DatabaseAvailabilityType.isEmbedded()" because "databaseAvailabilityType" is null
The field `datalakeDatabaseAvailabilityType` got introduced in Sep 26, 2023 with this commit:
https://github.infra.cloudera.com/cloudbreak/cloudbreak/commit/070b6f26f736db682ecbd3cb5d3e972f96662e7e

Entity fetch results in non-null field, because of the default converter.
This could be the exact cause of the NPE:

   Database database = stackDtoRepository.getDatabaseByStackById(stackView.getId()).orElse(new Database());

Unfortunately, it seems there was no data migration for the already running clusters, leaving their new field empty.
E.g. there are 133 affected rows in mow-dev, all before Sep 26, 2023.

This PR only handles the issue gracefully, without an NPE, proper solution would be a data migration, joining CB and SDX databases.

It could be covered with a stack patcher later.