CB-29521 Adjust AWS LB naming to handle DL resize
During DL resize a new cluster is created with the same CRN as the old.
As we switched to CRN based LB naming, it caused a name collision during resize, so no new LB was created for the new DL.
After the new cluster was created, the old is deleted, and deletes the old LB, and leaves the new cluster without LB.

Solution is to cut off the end of the CRN, and replace with 4 random alphanumeric characters ensuring that new LB would be created,
even though the CRN is the same.