CB-29547 Readdd `TG` to AWS Loadbalancer targetgroup and listener name matching
There could be a collision in case of Oozie where we expose 11443 and 443 port too.
The name would look like:
- mmoln-443Internal-5ad7b543-Y1Gf
- mmo-11443Internal-5ad7b543-Y1Gf

We moved back to:
- mmo-TG443Internal-5ad7b543-Y1Gf
- m-TG11443Internal-5ad7b543-Y1Gf

And the current logic would check for `443Internal`
Now the `TG` is added back so check would look for `TG443Internal` to avoid collision with `11443Internal`.

Checking for `-443Internal` wouldn't work with older clusters.