CB-29636 Move back `freeipa.<domain>` from LB to `ipa-ca` CNAME
`freeipa.<domain>` is used by tgtgen to communicate with FreeIPA through it's API on port 443.
As this is a stateful API, it cannot be loadbalanced the same way as the other ports.
To avoid introducing sticky loadbalancing everywhere for port 443, for now the `freeipa` CNAME would point to `ipa-ca` again.