CB-29683 Handle Freeipa optimistic lock exception during status update
Rare cases a flow and status checker can try to update the status concurrently.
To mitigate this in the affected flow the status updated is brought to the beginning of the method.
Also in case of an collision, the status update would be retried with refetching the stack, which should handle this.