CB-29721 Disable password change in SSSD in case of Active Directory
During install we copy the contet of /etc/krb5.keytab to cmf to use but if the keytab changes then CM won’t be able to generate service principals.
It changes because by default in case of AD, the computer password is changed every 30 days. The password change invalidates the keytab.
The simplest solution is to disable this password change, so the keytab will stay valid.

Doc:
https://docs.redhat.com/en/documentation/red_hat_enterprise_linux/7/html/windows_integration_guide/sssd-auto-keytab-renewal
https://jhrozek.fedorapeople.org/sssd/1.13.4/man/sssd-ad.5.html
https://learn.microsoft.com/en-us/previous-versions/windows/it-pro/windows-10/security/threat-protection/security-policy-settings/domain-member-maximum-machine-account-password-age