CB-29562: Clean PTR and A record from FreeIPA if the DNS A record already exists with other IP
In this commit I've improved the behaviour of the addDnsARecord endpoint.

The issue occurred when we attempted to add the A record for the new load balancer after the DL resize.
The DNS A record was already registered in FreeIPA with a different IP address. After deleting the old A record and attempting to add the new one, FreeIPA returned an error because the corresponding PTR record had not been removed along with the A record.

Therefore I made some changes to delete the A record and PTR record as well.