CB-29915 Configure the Data Hub to forward the DNS requests to the AD domain
In case of hybrid env, when configuring unbound, a new domain is added to the forwarders.
It's based on the FreeIPA trust response's realm field. It will forward these to the FreeIPA too,
as it has the forward zone already setup, this way it can be managed at one place.

```
$ cat /etc/unbound/conf.d/60-domain-dns.conf

forward-zone:
  name: "hybrid.cloudera.org"
  forward-addr: 10.84.137.147
  forward-addr: 10.84.165.194
forward-zone:
  name: "mmhybrid.xcu2-8y8x.wl.cloudera.site"
  forward-addr: 10.84.137.147
  forward-addr: 10.84.165.194
```