ENGESC-30628: Add validation step to the AddVolumesFlow
Added an extra validation step to the beginning of the AddVolumesFlow, which checks if all the instances in the group have the same number of attached volumes as we expect them to have based on the template. If the actual count of attached volumes is different than the expected, the flow is stopped.