CB-29898: Add validation for SELinux in E2E tests
The SELinux validations can run at the end of each E2E test. Currently this is turned off. The validations can also be used in the individual tests through the `SELinuxAssertions` class.

We are able to validate the SELinux mode, the number of installed custom SELinux policy modules created by us and the presence of denies in the audit log.
For the denies, the assertion returns the audit logs in question, and also pipes them into `audit2allow`, which generates allow rules which could be used to fix the given denies.