CB-1: Refactor `HostUtil.hasPort` to handle all our common edge cases
It tries to parse the input string as an URL, which works for most cases, expect if there is no protocol specified.
For inputs without a protocol, like `localhost:10090`, we fall back to a regex based approach, which tries to find the first colon (`:`) followed by 1 to 5 numbers, and then also checks that the next character is not a number, so that we don't end up matching 6 character long port nuumbers.