CB-30046: Do not stop services before Cloudera Manager upgrade
In this commit, I removed the code that stops all services before the Cloudera Manager upgrade.

This change is necessary because we encountered a scenario where the upgrade did not complete successfully. Specifically, when upgrading to a new image that only included a Cloudera Manager patch upgrade, the services were stopped, and we expected them to be restarted during the runtime upgrade phase, which did not happen as intended.