CB-30069 Optimize cleanup to handle large number of zones
In case of hybrid fetch DNS records for zones in batch, as in case of hybrid there can be hundreds of zones.
Each zone should have a single record at least, the NS record, which is used to fetch zone information when parsing response.
Zones with only a single record thus are excluded from cleanup, as those are NS records, which are not cleaned up.

Also the cleanup has been modified to fetch records in a zone only once.

Batch calls has been modified to handle response values.