CB-29843: Fix the `recipe-runner.sh`'s file context before it runs the recipes
This is a workaround needed because the file is created first under a temporary name with a random string at the end of the filename and then later renamed, so we can't create a type_transition rule for it, since knowing the exact final filename is needed for that.
We check if the `restorecon` command is available, and only run the command if it is. We also pass the `-i` flag to restorecon, which will stop the command from failing if the file, passed as the argument, does not exist.