CB-30266: Validate entitlement and image support for SELinux in FMS
Introduced the concept of tagFilters in image selection, which is a map of key value pairs, and checks if the image's tags have the specified filter keys with the same value as the filter. If the image's tags does not have the filter key, that filter key is ignored. So the absence of a filter key in the image's tags is treated the same as if it had the expected value, so we won't filter out images if they don't have a filter key. We will only filter them out if it explicitly has the filter key, and it has a different value.